+++ types-override/http/body.d.ts
@@ -1,5 +1,11 @@
-import type { BodyOptions, BodyResult } from '../ergo.js';
+import type {BodyOptions, BodyResult} from '../ergo.js';
 
-declare function body(options?: BodyOptions): (req: import('node:http').IncomingMessage) => Promise<BodyResult | { response: { statusCode: number; detail: string } }>;
+declare function body(
+  options?: BodyOptions
+): ((
+  req: import('node:http').IncomingMessage
+) => Promise<BodyResult | {response: {statusCode: number; detail: string}}>) & {
+  readonly setPath: 'body';
+};
 
 export default body;