+++ types-override/http/prefer.d.ts
@@ -1,5 +1,7 @@
-import type { PreferResult } from '../ergo.js';
+import type {PreferResult} from '../ergo.js';
 
-declare function prefer(): (req: import('node:http').IncomingMessage) => PreferResult;
+declare function prefer(): ((req: import('node:http').IncomingMessage) => PreferResult) & {
+  readonly setPath: 'prefer';
+};
 
 export default prefer;